requirements.txt
# Core dependencies
openai>=1.0.0
pymupdf>=1.23.0
rank-bm25>=0.2.2
numpy>=1.24.0
python-dotenv>=1.0.0

# Optional: for semantic search
faiss-cpu>=1.7.4
